segen: add static IP for segen-emessage
There is now a Raspberry Pi in the top level of Segen which
performes multiple tasks.  First, it has a Nooelec NESDR SMArt v4 SDR
to receive emessages.  Second. it is set to AP mode to provide WiFi
in the top level of Segen (the reception up to now was not
good, and the new addition makes it easier to do maintenance on the
top level).

Signed-off-by: pmelange <isprotejesvalkata@gmail.com>