template:tunspace: add deterministic mac
This behavior was already introduced in another commit. It is needed because the khsb locations are
behind a firewall that has an allow list for certain MACs. Before these changes, the MAC got randomly
generated by tunspace.
With the fixes, we have a determenitstic mac to explicitly allow this MAC in their network.